app/backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
pymongo==3.12.0
happybase==1.2.0
numpy==1.21.0
pandas==1.3.0
kafka-python==2.0.2
thriftpy2==0.4.16
PyJWT==2.3.0
bcrypt==3.2.0
python-dotenv==0.19.0